fix delay

+++ Helper/AppHelper.swift
@@ -202,8 +202,8 @@ func deepCopy<T: Codable>(_ object: T) -> T? {
 
 /// 延时
 @available(iOS 13.0, *)
-func delay(_ duration: UInt64) async {
-    try? await Task.sleep(nanoseconds: duration * NSEC_PER_SEC)
+func delay(_ duration: Double) async {
+    try? await Task.sleep(nanoseconds: UInt64(duration * Double(NSEC_PER_SEC)))
 }
 
 /// 打印调试信息
